requirements.txt
# pip install -r requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

accelerate
aiohttp
aiosignal
annotated-types
anyio
attrs
beautifulsoup4
bs4
certifi
charset-normalizer
click
dataclasses-json
Deprecated
dirtyjson
diskcache
distro
fastapi
filelock
frozenlist
fsspec
greenlet
guidance
h11
html2text
httpcore
httpx
huggingface-hub
idna
interegular
Jinja2
joblib
llama-hub
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-embeddings-instructor
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-huggingface
llama-index-llms-llama-cpp
llama-index-llms-ollama
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-guidance
llama-index-program-lmformatenforcer
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
llama_cpp_python
llamaindex-py-client
lm-format-enforcer
MarkupSafe
marshmallow
minijinja
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
openai
ordered-set
packaging
pandas
pillow
platformdirs
protobuf
psutil
pyaml
pydantic
pydantic_core
pydot
pyformlang
PyMuPDF
PyMuPDFb
pyparsing
pypdf
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
retrying
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
soupsieve
SQLAlchemy
starlette
striprtf
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tqdm
transformers
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wrapt
yarl
pydantic